% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ENRanalyticsS3.R
\docType{package}
\name{ENRanalyticsS3}
\alias{ENRanalyticsS3}
\alias{ENRanalyticsS3-package}
\title{ENRanalyticsS3 pakiet do ściagania danych z S3}
\description{
Tu bedzie opis pakietu
}
\section{Opis funkcji}{

hello2 funkcja testuje testthat \code{\link{hello2}}
}

