% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{set_db_conn}
\alias{set_db_conn}
\title{Set connection to redshift database.}
\usage{
set_db_conn(host_ip = "10.201.1.136", port = 5439,
  db_user = Sys.getenv("REVEL_DB_USER"),
  db_pass = Sys.getenv("REVEL_DB_PASS"), db_name = "cds")
}
\arguments{
\item{host_ip}{Ip adress of database server.}

\item{port}{Port number.}

\item{db_user}{A name of the user account. By default get from .Renviron file.}

\item{db_pass}{A password to the user account. By default get from .Renviron file.}

\item{db_name}{A name of the database.}
}
\value{
Connection string.
}
\description{
Set connection to redshift database.
}
